+++ detection.h
@@ -8,6 +8,6 @@
 void guess_mime_types(source_buffer *src);
 void get_snippet(source_buffer *src);
 void get_thumbnail(source_buffer *src);
-int find_write_type(source_buffer *src);
+uint8_t find_write_type(source_buffer *src);
 
 #endif
